billboarding

+++ PointsOfInterest/src/uk/ac/brookes/tederiksson/pointsofinterest/PoiPart.java
@@ -9,9 +9,9 @@ public class PoiPart implements Serializable {
 	private static final long serialVersionUID = -4862453600761412990L;
 	private int id;
 	private String message;
-	public Double x,y,z;
+	public float x,y,z;
 	
-	public PoiPart(int id, String message, Double x, Double y, Double z) {
+	public PoiPart(int id, String message, float x, float y, float z) {
 		super();
 		this.id = id;
 		this.message = message;
